Gestione test con h2
--HG--
rename : portofino-war/src/main/java/com/manydesigns/portofino/database/platforms/PostgreSQLDatabasePlatform.java => portofino-war/src/main/java/com/manydesigns/portofino/database/platforms/H2DatabasePlatform.java
rename : portofino-war/src/main/resources/databases/hibernatetest.sql => portofino-war/src/test/resources/database/hibernatetest.sql
rename : portofino-war/src/main/resources/databases/jpetstore/postgresql/jpetstore-postgres-dataload.sql => portofino-war/src/test/resources/database/jpetstore-postgres-dataload.sql
rename : portofino-war/src/main/resources/databases/jpetstore/postgresql/jpetstore-postgres-schema.sql => portofino-war/src/test/resources/database/jpetstore-postgres-schema.sql
rename : portofino-war/src/main/resources/portofino-connections.xml => portofino-war/src/test/resources/database/portofino-connections.xml
rename : portofino-war/src/main/resources/databases/portofino4.sql => portofino-war/src/test/resources/database/portofino4.sql

+++ portofino-war/src/main/java/com/manydesigns/portofino/system/model/users/Group.java
@@ -39,7 +39,7 @@ import java.util.List;
 * @author Giampiero Granatella - giampiero.granatella@manydesigns.com
 */
 public class Group implements Serializable{
-    Integer groupId;
+    Long groupId;
     String name;
     String description;
     Long creatorId;
@@ -49,11 +49,11 @@ public class Group implements Serializable{
     
     List<User> users = new ArrayList<User>();
 
-    public Integer getGroupId() {
+    public Long getGroupId() {
         return groupId;
     }
 
-    public void setGroupId(Integer groupId) {
+    public void setGroupId(Long groupId) {
         this.groupId = groupId;
     }
